+++ sinner/gui/controls/ThumbnailWidget/SourcesThumbnailWidget.py
@@ -1,4 +1,4 @@
-from typing import Tuple, Callable
+from typing import Tuple, Callable, Optional
 
 from PIL import Image
 
@@ -6,7 +6,7 @@ from sinner.gui.controls.ThumbnailWidget.BaseThumbnailWidget import BaseThumbnai
 from sinner.utilities import is_image
 
 
-class ImageThumbnailWidget(BaseThumbnailWidget):
+class SourcesThumbnailWidget(BaseThumbnailWidget):
 
     def add_thumbnail(self, source_path: str, caption: str | bool = True, click_callback: Callable[[str], None] | None = None) -> None:
         """
@@ -18,12 +18,12 @@ class ImageThumbnailWidget(BaseThumbnailWidget):
         if is_image(source_path):
             super().add_thumbnail(source_path, caption, click_callback)
 
-    def _prepare_thumbnail_data(self, source_path: str, caption: str | bool, click_callback: Callable[[str], None] | None) -> Tuple[Image.Image, str, str | bool, Callable[[str], None] | None]:
+    def _prepare_thumbnail_data(self, source_path: str, caption: str | bool, click_callback: Callable[[str], None] | None) -> Optional[Tuple[Image.Image, str, str | bool, Callable[[str], None]| None]]:
         """
         Prepare thumbnail data in background thread
         """
-        img = self.get_cached_thumbnail(source_path)
-        if not img:
-            img = self.get_thumbnail(Image.open(source_path), self.thumbnail_size)
-            self.set_cached_thumbnail(source_path, img)
-        return img, source_path, caption, click_callback
+        thumbnail = self.get_cached_thumbnail(source_path)
+        if not thumbnail:
+            thumbnail = self.get_thumbnail(Image.open(source_path), self.thumbnail_size)
+            self.set_cached_thumbnail(source_path, thumbnail)
+        return thumbnail, source_path, caption, click_callback
